Rename genericauth to simple
